Fixed performance and MMU creation
The performance issue was fixed by both decreasing the number of calls
to MMU::readByte() and fixing the number of cycles reported to
GPU::step().

Before this commit, MMU creation allowed for allocation of an array of
size 0 by passing in a value of 0 ram banks. This has been changed to
make eram a null pointer in this case and to check for it when reading
it.