Implemented Save States
Save states are now implemented, just click on file/Save State or Load
State to save or load a state.

To save and load states, the emulation is paused at the end of a VBLANK
period to write the contents of memory and CPU registers to disk. Right
now, the GPU isn't written to disk, but I have not encountered any
problems due to this so far. The only side-effect would be that mode
switches wouldn't be occurring at the exact right time.
Another possible issue with this is the lack of use of atomic
operations/variables in creating the mutex lock. This could be an issue,
but it is unlikely because a VBLANK is going to be emulated in ~2-3ms;
this is orders of magnitude less time needed for the user to select a
file to save/load to.

Also, exiting now works correctly.