Save State error checking and Bugfixes
Implemented a rudimentary way of checking save states which may change
in the future. Right now, a set of 10 bytes are written to the beginning
of the file and checked when loaded.

Also, fixed a bug where if the window was being displayed, then the BG
and window's first 8 pixels were not being drawn.

Finally, the screen is resizable now. I plan to fix it to a certain
aspect ratio if possible.