Header clean-up and Save state changes
Just cleaned up the header files, some of which were becoming a mess to
find things.

Also, changed the hash written to the beginning of a save state. Now,
the signature was expanded to 15 bytes and is xor'd with the ROM's name.
This attempts to limit the ability to load a save state with from a
different game.