Halt handling fixed
Fixed how halts were being handled. Before, they went to a separate
function that was similar to the UpdateToVBlank function, without
executing opcodes. However, this meant that there would be an occasional
second VBlank in a ~16ms window, doubling the speed at which frames were
updated. This was apparent in the uneven speed of emulation. Now, the
CPU's Update function just returns 4 cycles.

This correction has made a couple of things very apparent, however.
First, the wxTimer seemed to trigger at twice the requested amount of
time. Halving the time to 8 gives us the desired ~16ms update interval
we need, if not a little shorter. Also, the image scaling used at the
moment causes a large performance hit. I'll be looking at alternative
ways of accomplishing these tasks.