Timing Done
Finally fixed the issue with timing. I moved away from the wxTimer event
and am now using the wxIdle event to queue events quickly. As the events
come in, wxStopwatch provides a microsecond resolution counter to see if
enough time has passed. This allows it to update at the exactly desired
59.7fps +/- .01.