Fixed OptionFrame and implemented saving Cart RAM
Cart RAM is where most games save things like high scores or save games,
so there is now an option to save and load that. You could get around
ever needing this by using save states, but this is here as an
alternative.

OptionFrame was fixed in that closing it actually destroyed the object,
rather than just hiding the frame.